major restructuring of folders and files
