+++ base/_src/helpers.py
@@ -8,7 +8,7 @@ import sklearn.decomposition
 import sklearn.preprocessing
 from statsmodels.tsa.stattools import acf, pacf
 
-from spc.definitions import (
+from base._src.definitions import (
     LCL_color,
     UCL_color,
     get_line_width,
