+++ base/tests/test_barcharts.py
@@ -6,7 +6,7 @@ import pandas as pd
 import pytest
 from pandas._testing import assert_frame_equal, assert_series_equal
 
-from spc.core import EWMAChart, MEWMAChart, PCAModelChart, RChart, SChart, XBarChart
+from base import EWMAChart, MEWMAChart, PCAModelChart, RChart, SChart, XBarChart
 
 
 def sd(values):
