+++ base/tests/test_helpers.py
@@ -6,7 +6,7 @@ import pandas as pd
 import pytest
 from pandas._testing import assert_frame_equal
 
-from spc.helpers import (
+from base._src.helpers import (
     flatten_list,
     get_df_with_sample_id,
     get_num_of_PCs_to_retain,
