+++ dynamicPCA/analysis.py
